move training page
