+++ public/video-ui/src/pages/Training/index.js
@@ -19,6 +19,6 @@ export default class Training extends React.Component {
           </li>
         </ul>
       </div>
-    )
+    );
   }
 }
